Disable video tests with a flag.

+++ test/unit/video_processing_test.rb
@@ -4,11 +4,13 @@ class VideoProcessingTest < Test::Unit::TestCase
   include VideoProcessing
 
   def test_convert_video
-    video_dir = File.join(RAILS_ROOT, 'test', 'unit', 'data')
-    convert_video(File.join(video_dir, 'macbook.avi'))
-    assert File.exist?(File.join(video_dir, 'macbook.flv'))
-    File.delete(File.join(video_dir, 'macbook.flv'))
-    assert File.exist?(File.join(video_dir, 'macbook.jpg'))
-    File.delete(File.join(video_dir, 'macbook.jpg'))
+    if VIDEO_ENABLED
+      video_dir = File.join(RAILS_ROOT, 'test', 'unit', 'data')
+      convert_video(File.join(video_dir, 'macbook.avi'))
+      assert File.exist?(File.join(video_dir, 'macbook.flv'))
+      File.delete(File.join(video_dir, 'macbook.flv'))
+      assert File.exist?(File.join(video_dir, 'macbook.jpg'))
+      File.delete(File.join(video_dir, 'macbook.jpg'))
+    end
   end
 end
